requirements.txt
bcrypt==5.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
docutils==0.22.3
filetype==1.2.0
idna==3.11
Kivy==2.3.1
Kivy-Garden==0.1.5
kivy_deps.angle==0.4.0
kivy_deps.glew==0.3.1
kivy_deps.sdl2==0.8.0
kivymd==1.2.0
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pypiwin32==223
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0